+++ src/microhttpd/internal.h
@@ -782,13 +782,13 @@ struct MHD_Connection
    * otherwise, this is the size of the current chunk.  A value of
    * zero is also used when we're at the end of the chunks.
    */
-  unsigned int current_chunk_size;
+  size_t current_chunk_size;
 
   /**
    * If we are receiving with chunked encoding, where are we currently
    * with respect to the current chunk (at what offset / position)?
    */
-  unsigned int current_chunk_offset;
+  size_t current_chunk_offset;
 
   /**
    * Handler used for processing read connection operations
@@ -1217,7 +1217,7 @@ struct MHD_Daemon
   /**
    * Size of `digest_auth_random.
    */
-  unsigned int digest_auth_rand_size;
+  size_t digest_auth_rand_size;
 
   /**
    * Size of the nonce-nc array.
